+++ src/refactoring/hooks/useProductEditable.ts
@@ -1,0 +1,117 @@
+import { useState } from 'react';
+import { Discount, Product } from '../../types';
+
+type UpdatedKeyType = keyof Discount;
+
+interface IUseProductEditable {
+  products: Product[];
+  onProductUpdate: (updatedProduct: Product) => void;
+}
+
+export const useProductEditable = ({ products, onProductUpdate }: IUseProductEditable) => {
+  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
+  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
+  const [newDiscount, setNewDiscount] = useState<Discount>({ quantity: 0, rate: 0 });
+
+  const toggleProductAccordion = (productId: string) => {
+    setOpenProductIds((prev) => {
+      const newSet = new Set(prev);
+      if (newSet.has(productId)) {
+        newSet.delete(productId);
+      } else {
+        newSet.add(productId);
+      }
+      return newSet;
+    });
+  };
+
+  // handleEditProduct 함수 수정
+  const handleEditProduct = (product: Product) => {
+    setEditingProduct({ ...product });
+  };
+
+  // 새로운 핸들러 함수 추가
+  const handleProductNameUpdate = (productId: string, newName: string) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const updatedProduct = { ...editingProduct, name: newName };
+      setEditingProduct(updatedProduct);
+    }
+  };
+
+  // 새로운 핸들러 함수 추가
+  const handlePriceUpdate = (productId: string, newPrice: number) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const updatedProduct = { ...editingProduct, price: newPrice };
+      setEditingProduct(updatedProduct);
+    }
+  };
+
+  // 수정 완료 핸들러 함수 추가
+  const handleEditComplete = () => {
+    if (editingProduct) {
+      onProductUpdate(editingProduct);
+      setEditingProduct(null);
+    }
+  };
+
+  const handleStockUpdate = (productId: string, newStock: number) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct) {
+      const newProduct = { ...updatedProduct, stock: newStock };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+    }
+  };
+
+  const handleAddDiscount = (productId: string) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct && editingProduct) {
+      const newProduct = {
+        ...updatedProduct,
+        discounts: [...updatedProduct.discounts, newDiscount],
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+      setNewDiscount({ quantity: 0, rate: 0 });
+    }
+  };
+
+  const handleRemoveDiscount = (productId: string, index: number) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct) {
+      const newProduct = {
+        ...updatedProduct,
+        discounts: updatedProduct.discounts.filter((_, i) => i !== index),
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+    }
+  };
+
+  const handleChangeDisCountForm =
+    (updatedKey: UpdatedKeyType) => (e: React.ChangeEvent<HTMLInputElement>) =>
+      setNewDiscount({
+        ...newDiscount,
+        [updatedKey]: getUpdatedValue(updatedKey, e.target.value),
+      });
+
+  const getUpdatedValue = (key: UpdatedKeyType, value: string) => {
+    return key === 'quantity' ? parseInt(value) : parseInt(value) / 100;
+  };
+
+  return {
+    openProductIds,
+    editingProduct,
+    newDiscount,
+    setNewDiscount,
+    toggleProductAccordion,
+    handleEditProduct,
+    handleProductNameUpdate,
+    handlePriceUpdate,
+    handleEditComplete,
+    handleStockUpdate,
+    handleAddDiscount,
+    handleChangeDisCountForm,
+    handleRemoveDiscount,
+  };
+};
